pyproject.toml
[project]
name = "bot-testing"
version = "0.1.0"
authors = [
    {name = "Helen Dobruk", email = "elena.dobruk@yandex.ru"},
]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.18.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "flake8>=7.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
